user/requirements.txt
bcrypt==3.2.0
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
cryptography==3.2.1
ecdsa==0.14.1
fastapi==0.61.1
fastapi-utils==0.2.1
gunicorn==20.0.4
h11==0.11.0
idna==2.10
mongoengine==0.20.0
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.7.2
PyJWT==1.7.1
pymongo==3.11.0
python-jose==3.2.0
python-multipart==0.0.5
requests==2.24.0
rsa==4.6
six==1.15.0
SQLAlchemy==1.3.20
starlette==0.13.6
urllib3==1.25.11
uvicorn==0.12.2
pyotp==2.4.1
jinja2